:recycle: refactor: 파일 확장자 일괄 변경, createRoot 및 App 분리
